refactor(handlers): slim FileHandler deps, split file.go by seam
Drop the now-dead faceSvc/objSvc dependencies (NewFileHandler 11->9 params) and
split the 1047-line file.go into three cohesive files: file.go (CRUD + shared
helpers), file_serve.go (streaming/proxy/thumbnail serving), file_jobs.go
(media-job triggers + bulk). Pure relocation + dependency-arity change; no
behavior change.
