+++ backend/internal/handlers/file.go
@@ -2,12 +2,8 @@ package handlers
 
 import (
 	"errors"
-	"fmt"
 	"io"
 	"net/http"
-	"regexp"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
@@ -18,12 +14,10 @@ import (
 	"github.com/alcoves/alcoves-backend/internal/models"
 	"github.com/alcoves/alcoves-backend/internal/services/activity"
 	"github.com/alcoves/alcoves-backend/internal/services/audiodetection"
-	"github.com/alcoves/alcoves-backend/internal/services/facedetection"
 	"github.com/alcoves/alcoves-backend/internal/services/filehash"
 	"github.com/alcoves/alcoves-backend/internal/services/files"
 	"github.com/alcoves/alcoves-backend/internal/services/mediajobs"
 	"github.com/alcoves/alcoves-backend/internal/services/metadata"
-	"github.com/alcoves/alcoves-backend/internal/services/objectdetection"
 	"github.com/alcoves/alcoves-backend/internal/services/storage"
 	"github.com/alcoves/alcoves-backend/internal/services/transcribe"
 	"github.com/alcoves/alcoves-backend/internal/services/videoproxy"
@@ -34,8 +28,6 @@ type FileHandler struct {
 	db             *gorm.DB
 	fileSvc        *files.Service
 	storageSvc     *storage.Service
-	faceSvc        *facedetection.Service
-	objSvc         *objectdetection.Service
 	videoSvc       *videoproxy.Service
 	transcribeSvc  *transcribe.Service
 	audioDetectSvc *audiodetection.Service
@@ -45,8 +37,8 @@ type FileHandler struct {
 	mediaJobs      *mediajobs.Service
 }
 
-func NewFileHandler(db *gorm.DB, fileSvc *files.Service, storageSvc *storage.Service, faceSvc *facedetection.Service, objSvc *objectdetection.Service, videoSvc *videoproxy.Service, transcribeSvc *transcribe.Service, audioDetectSvc *audiodetection.Service, waveformSvc *waveform.Service, metadataSvc *metadata.Service, activitySvc *activity.Service) *FileHandler {
-	return &FileHandler{db: db, fileSvc: fileSvc, storageSvc: storageSvc, faceSvc: faceSvc, objSvc: objSvc, videoSvc: videoSvc, transcribeSvc: transcribeSvc, audioDetectSvc: audioDetectSvc, waveformSvc: waveformSvc, metadataSvc: metadataSvc, activitySvc: activitySvc, mediaJobs: mediajobs.NewService(db, videoSvc, waveformSvc, transcribeSvc, audioDetectSvc)}
+func NewFileHandler(db *gorm.DB, fileSvc *files.Service, storageSvc *storage.Service, videoSvc *videoproxy.Service, transcribeSvc *transcribe.Service, audioDetectSvc *audiodetection.Service, waveformSvc *waveform.Service, metadataSvc *metadata.Service, activitySvc *activity.Service) *FileHandler {
+	return &FileHandler{db: db, fileSvc: fileSvc, storageSvc: storageSvc, videoSvc: videoSvc, transcribeSvc: transcribeSvc, audioDetectSvc: audioDetectSvc, waveformSvc: waveformSvc, metadataSvc: metadataSvc, activitySvc: activitySvc, mediaJobs: mediajobs.NewService(db, videoSvc, waveformSvc, transcribeSvc, audioDetectSvc)}
 }
 
 func (h *FileHandler) RegisterRoutes(g *echo.Group) {
@@ -162,73 +154,6 @@ func (h *FileHandler) Get(c echo.Context) error {
 		dupes, _ = filehash.FindDuplicates(h.db, file.LibraryID, file.ID, *file.Hash)
 	}
 	return c.JSON(http.StatusOK, fileToJSON(file, dupes))
-}
-
-func (h *FileHandler) serveFileData(c echo.Context, file *models.File) error {
-	libraryID := file.LibraryID.String()
-	fileID := file.ID.String()
-
-	size, err := h.storageSvc.FileStat(libraryID, fileID)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "File not found on storage")
-	}
-
-	c.Response().Header().Set("Content-Type", file.MimeType)
-	c.Response().Header().Set("Accept-Ranges", "bytes")
-	c.Response().Header().Set("Content-Disposition", fmt.Sprintf(`inline; filename="%s"`, file.Name))
-
-	rangeHeader := c.Request().Header.Get("Range")
-	if rangeHeader != "" {
-		return h.serveRangeRequest(c, libraryID, fileID, size, rangeHeader)
-	}
-
-	reader, err := h.storageSvc.OpenFileReadStream(libraryID, fileID, nil)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to read file")
-	}
-	defer reader.Close()
-
-	c.Response().Header().Set("Content-Length", strconv.FormatInt(size, 10))
-	return c.Stream(http.StatusOK, file.MimeType, reader)
-}
-
-var rangeRegex = regexp.MustCompile(`bytes=(\d+)-(\d*)`)
-
-func (h *FileHandler) serveRangeRequest(c echo.Context, libraryID, fileID string, totalSize int64, rangeHeader string) error {
-	matches := rangeRegex.FindStringSubmatch(rangeHeader)
-	if matches == nil {
-		return echo.NewHTTPError(http.StatusRequestedRangeNotSatisfiable, "Invalid range")
-	}
-
-	start, _ := strconv.ParseInt(matches[1], 10, 64)
-	var end int64
-	if matches[2] != "" {
-		end, _ = strconv.ParseInt(matches[2], 10, 64)
-	} else {
-		end = totalSize - 1
-	}
-
-	if start >= totalSize {
-		c.Response().Header().Set("Content-Range", fmt.Sprintf("bytes */%d", totalSize))
-		return c.NoContent(http.StatusRequestedRangeNotSatisfiable)
-	}
-
-	if end >= totalSize {
-		end = totalSize - 1
-	}
-
-	length := end - start + 1
-	reader, err := h.storageSvc.OpenFileReadStream(libraryID, fileID, &storage.ByteRange{Start: start, End: end})
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to read file")
-	}
-	defer reader.Close()
-
-	c.Response().Header().Set("Content-Range", fmt.Sprintf("bytes %d-%d/%d", start, end, totalSize))
-	c.Response().Header().Set("Content-Length", strconv.FormatInt(length, 10))
-	c.Response().WriteHeader(http.StatusPartialContent)
-	io.Copy(c.Response(), reader)
-	return nil
 }
 
 type updateFileRequest struct {
@@ -391,47 +316,6 @@ func (h *FileHandler) Restore(c echo.Context) error {
 	return c.JSON(http.StatusOK, map[string]int64{"restored": result.RowsAffected})
 }
 
-func (h *FileHandler) ReprocessVideoThumbnails(c echo.Context) error {
-	if h.videoSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Video service unavailable")
-	}
-
-	la := middleware.GetLibraryAccess(c)
-	if la == nil || !la.IsOwner {
-		return echo.NewHTTPError(http.StatusForbidden, "Only the library owner can regenerate video thumbnails")
-	}
-
-	libraryID := c.Param("id")
-	queuedCount, err := h.videoSvc.EnqueueExistingVideoThumbnails(libraryID)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to queue video thumbnails")
-	}
-
-	return c.JSON(http.StatusOK, map[string]int{"queuedCount": queuedCount})
-}
-
-// MetadataReprocess re-enqueues EXIF/media-metadata extraction for every media
-// file in the library. Owner/admin-gated by the route middleware (non-GET on
-// /api/libraries/:id/* requires library admin). Resets the 3-strike attempt cap
-// so previously-exhausted files retry — the escape hatch after a parser fix.
-func (h *FileHandler) MetadataReprocess(c echo.Context) error {
-	if h.metadataSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Metadata service unavailable")
-	}
-
-	libraryID := c.Param("id")
-	if _, err := uuid.Parse(libraryID); err != nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "Invalid library ID")
-	}
-
-	queuedCount, err := h.metadataSvc.ReprocessLibrary(libraryID)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, fmt.Sprintf("Reprocess failed: %v", err))
-	}
-
-	return c.JSON(http.StatusOK, map[string]int{"queuedCount": queuedCount})
-}
-
 // Timeline returns library files flattened and sorted newest-first by effective
 // capture date for the Timeline view. ?type=media (default) limits to images +
 // videos; ?type=all includes every file. Cursor-paginated.
@@ -461,390 +345,6 @@ func (h *FileHandler) Map(c echo.Context) error {
 		return err
 	}
 	return c.JSON(http.StatusOK, result)
-}
-
-type playbackSourceResponse struct {
-	ID        string `json:"id"`
-	Name      string `json:"name"`
-	MimeType  string `json:"mimeType"`
-	Kind      string `json:"kind"`
-	StreamURL string `json:"streamUrl"`
-	CreatedAt string `json:"createdAt"`
-}
-
-type playbackSourcesResponse struct {
-	DefaultSourceID string                   `json:"defaultSourceId"`
-	Sources         []playbackSourceResponse `json:"sources"`
-}
-
-func (h *FileHandler) PlaybackSources(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	selected, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-
-	if !strings.HasPrefix(selected.MimeType, "video/") {
-		return echo.NewHTTPError(http.StatusBadRequest, "Playback sources are only available for video files")
-	}
-
-	sourceID := selected.ID
-	if selected.SourceFileID != nil {
-		sourceID = *selected.SourceFileID
-	}
-
-	var source models.File
-	if err := h.db.Where("id = ? AND library_id = ?", sourceID, libraryID).First(&source).Error; err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "Source file not found")
-	}
-
-	var proxies []models.File
-	if err := h.db.
-		Where("source_file_id = ? AND library_id = ? AND trashed_at IS NULL AND mime_type LIKE ?", sourceID, libraryID, "video/%").
-		Order("created_at DESC").
-		Find(&proxies).Error; err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to load playback sources")
-	}
-
-	defaultSourceID := source.ID.String()
-	if source.ProxyStatus != nil && *source.ProxyStatus == "ready" && len(proxies) > 0 {
-		defaultSourceID = proxies[0].ID.String()
-	}
-
-	sources := make([]playbackSourceResponse, 0, len(proxies)+1)
-	sources = append(sources, playbackSourceResponse{
-		ID:        source.ID.String(),
-		Name:      source.Name,
-		MimeType:  source.MimeType,
-		Kind:      "source",
-		StreamURL: fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, source.ID.String()),
-		CreatedAt: source.CreatedAt.Format(time.RFC3339Nano),
-	})
-	for _, proxy := range proxies {
-		sources = append(sources, playbackSourceResponse{
-			ID:        proxy.ID.String(),
-			Name:      proxy.Name,
-			MimeType:  proxy.MimeType,
-			Kind:      "proxy",
-			StreamURL: fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, proxy.ID.String()),
-			CreatedAt: proxy.CreatedAt.Format(time.RFC3339Nano),
-		})
-	}
-
-	if len(proxies) == 0 && source.ProxyStatus != nil && *source.ProxyStatus == "ready" {
-		cacheKey := storage.ProxyKey(libraryID, source.ID.String())
-		if exists, _ := h.storageSvc.CacheExists(cacheKey); exists {
-			defaultSourceID = source.ID.String() + "::legacy-proxy"
-			sources = append(sources, playbackSourceResponse{
-				ID:        defaultSourceID,
-				Name:      "Legacy Proxy",
-				MimeType:  "video/mp4",
-				Kind:      "proxy",
-				StreamURL: fmt.Sprintf("/api/libraries/%s/files/%s/proxy", libraryID, source.ID.String()),
-				CreatedAt: source.UpdatedAt.Format(time.RFC3339Nano),
-			})
-		}
-	}
-
-	return c.JSON(http.StatusOK, playbackSourcesResponse{DefaultSourceID: defaultSourceID, Sources: sources})
-}
-
-func (h *FileHandler) GenerateProxy(c echo.Context) error {
-	if h.videoSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Video proxy service unavailable")
-	}
-
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-
-	if !strings.HasPrefix(file.MimeType, "video/") {
-		return echo.NewHTTPError(http.StatusBadRequest, "File is not a video")
-	}
-	if file.SourceFileID != nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "Cannot generate proxy for proxy file")
-	}
-
-	if err := h.mediaJobs.TriggerProxy(libraryID, file); err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to queue video proxy generation")
-	}
-
-	return c.JSON(http.StatusOK, h.fileToJSONWithLookup(file))
-}
-
-func (h *FileHandler) GenerateWaveform(c echo.Context) error {
-	if h.waveformSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Waveform service unavailable")
-	}
-
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-
-	if !strings.HasPrefix(file.MimeType, "video/") && !strings.HasPrefix(file.MimeType, "audio/") {
-		return echo.NewHTTPError(http.StatusBadRequest, "File is not audio/video")
-	}
-	if file.SourceFileID != nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "Cannot generate waveform for proxy file")
-	}
-
-	if err := h.mediaJobs.TriggerWaveform(libraryID, file); err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to queue waveform generation")
-	}
-
-	return c.JSON(http.StatusAccepted, h.fileToJSONWithLookup(file))
-}
-
-func (h *FileHandler) GetWaveform(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-	if file.WaveformStatus == nil || *file.WaveformStatus != "ready" {
-		return echo.NewHTTPError(http.StatusNotFound, "Waveform not ready")
-	}
-
-	cacheKey := storage.WaveformKey(libraryID, fileID)
-	// Stream the waveform JSON without loading it fully into RAM.
-	rc, err := h.storageSvc.OpenCacheReadStream(cacheKey)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "Waveform data not found")
-	}
-	defer rc.Close()
-	return c.Stream(http.StatusOK, "application/json", rc)
-}
-
-func (h *FileHandler) GenerateTranscript(c echo.Context) error {
-	if h.transcribeSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Transcribe service unavailable")
-	}
-
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-
-	if !strings.HasPrefix(file.MimeType, "video/") && !strings.HasPrefix(file.MimeType, "audio/") {
-		return echo.NewHTTPError(http.StatusBadRequest, "File is not audio/video")
-	}
-	if file.SourceFileID != nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "Cannot transcribe proxy file")
-	}
-
-	if err := h.mediaJobs.TriggerTranscribe(libraryID, file); err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to queue transcription")
-	}
-
-	return c.JSON(http.StatusAccepted, h.fileToJSONWithLookup(file))
-}
-
-func (h *FileHandler) GetTranscript(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-	if file.TranscribeStatus == nil || *file.TranscribeStatus != "ready" {
-		return echo.NewHTTPError(http.StatusNotFound, "Transcript not ready")
-	}
-
-	resp := map[string]interface{}{
-		"text":  file.TranscriptText,
-		"vtt":   file.TranscriptVTT,
-		"model": file.TranscriptModel,
-	}
-	return c.JSON(http.StatusOK, resp)
-}
-
-func (h *FileHandler) GenerateAudioDetections(c echo.Context) error {
-	if h.audioDetectSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Audio detection service unavailable")
-	}
-
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	file, err := findActiveFile(h.db, libraryID, fileID)
-	if err != nil {
-		return err
-	}
-	if !strings.HasPrefix(file.MimeType, "video/") && !strings.HasPrefix(file.MimeType, "audio/") {
-		return echo.NewHTTPError(http.StatusBadRequest, "File is not audio/video")
-	}
-	if file.TranscribeStatus == nil || *file.TranscribeStatus != "ready" {
-		return echo.NewHTTPError(http.StatusBadRequest, "Transcription must be completed before audio detection")
-	}
-	if file.SourceFileID != nil {
-		return echo.NewHTTPError(http.StatusBadRequest, "Cannot run detection on proxy file")
-	}
-
-	if err := h.mediaJobs.TriggerAudioDetect(libraryID, file); err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to queue audio detection")
-	}
-
-	return c.JSON(http.StatusAccepted, h.fileToJSONWithLookup(file))
-}
-
-func (h *FileHandler) ListAudioDetections(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	var file models.File
-	if err := h.db.Select("id").Where("id = ? AND library_id = ? AND trashed_at IS NULL", fileID, libraryID).First(&file).Error; err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "File not found")
-	}
-
-	if h.audioDetectSvc == nil {
-		return c.JSON(http.StatusOK, []interface{}{})
-	}
-	dets, err := h.audioDetectSvc.ListByFile(libraryID, fileID)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to list detections")
-	}
-	return c.JSON(http.StatusOK, dets)
-}
-
-func (h *FileHandler) Proxy(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	var file models.File
-	if err := h.db.Select("id, mime_type, proxy_status, name, source_file_id").
-		Where("id = ? AND library_id = ?", fileID, libraryID).First(&file).Error; err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "File not found")
-	}
-
-	if file.SourceFileID != nil {
-		return c.Redirect(http.StatusFound, fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, file.ID.String()))
-	}
-
-	var proxyFile models.File
-	err := h.db.Select("id, mime_type").
-		Where("source_file_id = ? AND library_id = ? AND trashed_at IS NULL", file.ID, libraryID).
-		Order("created_at DESC").
-		First(&proxyFile).Error
-	if err == nil {
-		return c.Redirect(http.StatusFound, fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, proxyFile.ID.String()))
-	}
-
-	if file.ProxyStatus != nil && *file.ProxyStatus == "not_needed" {
-		return c.Redirect(http.StatusFound, fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, fileID))
-	}
-
-	if file.ProxyStatus == nil || *file.ProxyStatus != "ready" {
-		msg := "No proxy available"
-		if file.ProxyStatus != nil && (*file.ProxyStatus == "processing" || *file.ProxyStatus == "queued") {
-			msg = "Proxy is still processing"
-		}
-		return echo.NewHTTPError(http.StatusNotFound, msg)
-	}
-
-	cacheKey := storage.ProxyKey(libraryID, fileID)
-	exists, _ := h.storageSvc.CacheExists(cacheKey)
-	if !exists {
-		return echo.NewHTTPError(http.StatusNotFound, "Proxy file not found")
-	}
-
-	// Get total size without loading the file into memory.
-	totalSize, err := h.storageSvc.CacheStat(cacheKey)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to stat proxy")
-	}
-
-	c.Response().Header().Set("Content-Type", "video/mp4")
-	c.Response().Header().Set("Accept-Ranges", "bytes")
-	c.Response().Header().Set("Cache-Control", "private, max-age=3600")
-
-	rangeHeader := c.Request().Header.Get("Range")
-	if rangeHeader != "" {
-		matches := rangeRegex.FindStringSubmatch(rangeHeader)
-		if matches != nil {
-			start, _ := strconv.ParseInt(matches[1], 10, 64)
-			var end int64
-			if matches[2] != "" {
-				end, _ = strconv.ParseInt(matches[2], 10, 64)
-			} else {
-				end = totalSize - 1
-			}
-
-			if start >= totalSize {
-				c.Response().Header().Set("Content-Range", fmt.Sprintf("bytes */%d", totalSize))
-				return c.NoContent(http.StatusRequestedRangeNotSatisfiable)
-			}
-			if end >= totalSize {
-				end = totalSize - 1
-			}
-
-			rc, err := h.storageSvc.OpenCacheReadStreamRange(cacheKey, &storage.ByteRange{Start: start, End: end})
-			if err != nil {
-				return echo.NewHTTPError(http.StatusInternalServerError, "Failed to read proxy")
-			}
-			defer rc.Close()
-
-			c.Response().Header().Set("Content-Range", fmt.Sprintf("bytes %d-%d/%d", start, end, totalSize))
-			c.Response().Header().Set("Content-Length", strconv.FormatInt(end-start+1, 10))
-			return c.Stream(http.StatusPartialContent, "video/mp4", rc)
-		}
-	}
-
-	// Full (non-range) response — stream without buffering.
-	rc, err := h.storageSvc.OpenCacheReadStream(cacheKey)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to read proxy")
-	}
-	defer rc.Close()
-
-	c.Response().Header().Set("Content-Length", strconv.FormatInt(totalSize, 10))
-	return c.Stream(http.StatusOK, "video/mp4", rc)
-}
-
-func (h *FileHandler) Thumbnail(c echo.Context) error {
-	libraryID := c.Param("id")
-	fileID := c.Param("fileId")
-
-	// Check file exists
-	var file models.File
-	if err := h.db.Select("id, thumbnail_file_id").Where("id = ? AND library_id = ?", fileID, libraryID).First(&file).Error; err != nil {
-		return echo.NewHTTPError(http.StatusNotFound, "File not found")
-	}
-
-	if file.ThumbnailFileID != nil {
-		return c.Redirect(http.StatusFound, fmt.Sprintf("/api/libraries/%s/files/%s?inline=true", libraryID, file.ThumbnailFileID.String()))
-	}
-
-	cacheKey := storage.ThumbnailKey(libraryID, fileID)
-	exists, _ := h.storageSvc.CacheExists(cacheKey)
-	if !exists {
-		return echo.NewHTTPError(http.StatusNotFound, "Thumbnail not found")
-	}
-
-	// Stream the thumbnail without loading it fully into RAM.
-	rc, err := h.storageSvc.OpenCacheReadStream(cacheKey)
-	if err != nil {
-		return echo.NewHTTPError(http.StatusInternalServerError, "Failed to read thumbnail")
-	}
-	defer rc.Close()
-
-	c.Response().Header().Set("Cache-Control", "private, max-age=86400")
-	return c.Stream(http.StatusOK, "image/webp", rc)
 }
 
 // fileToJSONWithLookup serializes a File and computes duplicateOfFileIds via
@@ -935,113 +435,4 @@ func timeStr(t *time.Time) *string {
 	}
 	s := t.Format(time.RFC3339Nano)
 	return &s
-}
-
-// bulkActionRequest is the shared shape for bulk-transcribe / bulk-audio-detect.
-// Empty FileIDs means "every video/audio file in the library" (the
-// "transcribe-everything" library-page button).
-type bulkActionRequest struct {
-	FileIDs []string `json:"fileIds,omitempty"`
-}
-
-// bulkActionResponse reports which files were enqueued vs. skipped so the
-// frontend can toast a useful summary instead of a binary success/fail.
-type bulkActionResponse struct {
-	Enqueued []string          `json:"enqueued"`
-	Skipped  map[string]string `json:"skipped"`
-}
-
-// BulkTranscribe queues transcription for many files at once. With an empty
-// fileIds array it queues every video/audio file in the library that isn't
-// a proxy. Dedup is handled by asynq.Unique on the enqueue side, so a user
-// double-clicking the button cannot fan out duplicate workers.
-func (h *FileHandler) BulkTranscribe(c echo.Context) error {
-	if h.transcribeSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Transcribe service unavailable")
-	}
-	libraryID := c.Param("id")
-	files, err := h.bulkResolveFiles(c, libraryID)
-	if err != nil {
-		return err
-	}
-
-	resp := bulkActionResponse{Enqueued: []string{}, Skipped: map[string]string{}}
-	for _, f := range files {
-		fid := f.ID.String()
-		if f.SourceFileID != nil {
-			resp.Skipped[fid] = "proxy file"
-			continue
-		}
-		if !strings.HasPrefix(f.MimeType, "video/") && !strings.HasPrefix(f.MimeType, "audio/") {
-			resp.Skipped[fid] = "not audio/video"
-			continue
-		}
-		if err := h.mediaJobs.TriggerTranscribe(libraryID, &f); err != nil {
-			resp.Skipped[fid] = err.Error()
-			continue
-		}
-		resp.Enqueued = append(resp.Enqueued, fid)
-	}
-	return c.JSON(http.StatusAccepted, resp)
-}
-
-// BulkAudioDetect queues PANNs audio-event detection across many files.
-// Skips files whose transcript isn't ready (matches the per-file endpoint
-// invariant) so the bulk button on a freshly-imported library still
-// produces useful feedback instead of a wall of errors.
-func (h *FileHandler) BulkAudioDetect(c echo.Context) error {
-	if h.audioDetectSvc == nil {
-		return echo.NewHTTPError(http.StatusServiceUnavailable, "Audio detection service unavailable")
-	}
-	libraryID := c.Param("id")
-	files, err := h.bulkResolveFiles(c, libraryID)
-	if err != nil {
-		return err
-	}
-
-	resp := bulkActionResponse{Enqueued: []string{}, Skipped: map[string]string{}}
-	for _, f := range files {
-		fid := f.ID.String()
-		if f.SourceFileID != nil {
-			resp.Skipped[fid] = "proxy file"
-			continue
-		}
-		if !strings.HasPrefix(f.MimeType, "video/") && !strings.HasPrefix(f.MimeType, "audio/") {
-			resp.Skipped[fid] = "not audio/video"
-			continue
-		}
-		if f.TranscribeStatus == nil || *f.TranscribeStatus != "ready" {
-			resp.Skipped[fid] = "transcript not ready"
-			continue
-		}
-		if err := h.mediaJobs.TriggerAudioDetect(libraryID, &f); err != nil {
-			resp.Skipped[fid] = err.Error()
-			continue
-		}
-		resp.Enqueued = append(resp.Enqueued, fid)
-	}
-	return c.JSON(http.StatusAccepted, resp)
-}
-
-// bulkResolveFiles loads the target file set for a bulk action. With an
-// explicit fileIds list it scopes to those IDs (after enforcing they belong
-// to the library); otherwise it returns every non-trashed video/audio file
-// in the library that isn't a transcoded proxy.
-func (h *FileHandler) bulkResolveFiles(c echo.Context, libraryID string) ([]models.File, error) {
-	var req bulkActionRequest
-	if err := c.Bind(&req); err != nil {
-		return nil, echo.NewHTTPError(http.StatusBadRequest, "Invalid request body")
-	}
-
-	q := h.db.Where("library_id = ? AND trashed_at IS NULL AND source_file_id IS NULL", libraryID).
-		Where("mime_type LIKE 'video/%' OR mime_type LIKE 'audio/%'")
-	if len(req.FileIDs) > 0 {
-		q = q.Where("id IN ?", req.FileIDs)
-	}
-
-	var files []models.File
-	if err := q.Find(&files).Error; err != nil {
-		return nil, echo.NewHTTPError(http.StatusInternalServerError, "Failed to load files")
-	}
-	return files, nil
 }
