+++ backend/internal/handlers/file_purge_test.go
@@ -245,7 +245,7 @@ func assertRecordCount(t *testing.T, db *gorm.DB, model interface{}, where strin
 func TestPurge_SpecificTrashedFiles(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	f1 := createFile(t, db, fix.LibraryID, fix.UserID, "photo1.jpg", true, nil)
@@ -283,7 +283,7 @@ func TestPurge_SpecificTrashedFiles(t *testing.T) {
 func TestPurge_IgnoresNonTrashedFiles(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	active := createFile(t, db, fix.LibraryID, fix.UserID, "active.jpg", false, nil)
@@ -317,7 +317,7 @@ func TestPurge_IgnoresNonTrashedFiles(t *testing.T) {
 func TestPurge_IgnoresFilesFromOtherLibrary(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix1 := seedLibrary(t, db)
 	fix2 := seedLibrary(t, db)
@@ -345,7 +345,7 @@ func TestPurge_IgnoresFilesFromOtherLibrary(t *testing.T) {
 func TestPurge_DeletesBlobAndLegacyCache(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "cached.jpg", true, nil)
@@ -382,7 +382,7 @@ func TestPurge_DeletesBlobAndLegacyCache(t *testing.T) {
 func TestPurge_RemovesFileTags(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "tagged.jpg", true, nil)
@@ -407,7 +407,7 @@ func TestPurge_RemovesFileTags(t *testing.T) {
 func TestPurge_DeletesDerivedFilesWithSource(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	original := createFile(t, db, fix.LibraryID, fix.UserID, "original.mp4", true, nil)
@@ -476,7 +476,7 @@ func TestPurge_DeletesDerivedFilesWithSource(t *testing.T) {
 func TestPurge_AllTrashedItems(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 
@@ -526,7 +526,7 @@ func TestPurge_AllTrashedItems(t *testing.T) {
 func TestPurge_AllTrashedWithEmptyJSONBody(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "trash.jpg", true, nil)
@@ -548,7 +548,7 @@ func TestPurge_AllTrashedWithEmptyJSONBody(t *testing.T) {
 func TestPurge_AllTrashedNothingToPurge(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	// Only active items
@@ -573,7 +573,7 @@ func TestPurge_AllTrashedNothingToPurge(t *testing.T) {
 func TestPurge_SpecificFolder(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	folderID := createFolder(t, db, fix.LibraryID, "photos", true, nil)
@@ -600,7 +600,7 @@ func TestPurge_SpecificFolder(t *testing.T) {
 func TestPurge_FolderWithDescendants(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 
@@ -647,7 +647,7 @@ func TestPurge_FolderWithDescendants(t *testing.T) {
 func TestPurge_FolderRemovesFolderTags(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	folderID := createFolder(t, db, fix.LibraryID, "tagged-folder", true, nil)
@@ -669,7 +669,7 @@ func TestPurge_FolderRemovesFolderTags(t *testing.T) {
 func TestPurge_FolderPurgesFilesAndTheirTags(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	folderID := createFolder(t, db, fix.LibraryID, "folder", true, nil)
@@ -716,7 +716,7 @@ func (d *failingDeleteDriver) DeletePrefix(scope storage.Scope, _ string) error 
 func TestPurge_StorageDeleteFailsReturns500(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	storageSvc := storage.NewService(&failingDeleteDriver{})
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: storageSvc}), storageSvc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: storageSvc}), storageSvc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "fail.jpg", true, nil)
@@ -739,7 +739,7 @@ func TestPurge_StorageDeleteFailsReturns500(t *testing.T) {
 func TestPurge_StorageDeleteFailsPreservesDBRecords(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	storageSvc := storage.NewService(&failingDeleteDriver{})
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: storageSvc}), storageSvc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: storageSvc}), storageSvc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "fail.jpg", true, nil)
@@ -757,7 +757,7 @@ func TestPurge_StorageDeleteFailsPreservesDBRecords(t *testing.T) {
 func TestPurge_InvalidBodyReturnsBadRequest(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 
@@ -790,7 +790,7 @@ func TestPurge_InvalidBodyReturnsBadRequest(t *testing.T) {
 func TestPurge_EmptyFileIDsListPurgesAllTrashed(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "only-trashed.jpg", true, nil)
@@ -813,7 +813,7 @@ func TestPurge_EmptyFileIDsListPurgesAllTrashed(t *testing.T) {
 func TestPurge_NonexistentFileIDs(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fakeID := uuid.New().String()
@@ -834,7 +834,7 @@ func TestPurge_NonexistentFileIDs(t *testing.T) {
 func TestPurge_MixedTrashedAndActiveInBulk(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	trashed1 := createFile(t, db, fix.LibraryID, fix.UserID, "t1.jpg", true, nil)
@@ -874,7 +874,7 @@ func TestPurge_MixedTrashedAndActiveInBulk(t *testing.T) {
 func TestPurge_MultipleFilesWithTags(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	sharedTag := createTag(t, db, fix.LibraryID, "shared")
@@ -906,7 +906,7 @@ func TestPurge_MultipleFilesWithTags(t *testing.T) {
 func TestPurge_FolderOnlyNoFilesInside(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	folderID := createFolder(t, db, fix.LibraryID, "empty-folder", true, nil)
@@ -929,7 +929,7 @@ func TestPurge_FolderOnlyNoFilesInside(t *testing.T) {
 func TestPurge_ResponseFormat(t *testing.T) {
 	db := setupPurgeTestDB(t)
 	svc := setupPurgeStorage(t)
-	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil, nil, nil)
+	handler := NewFileHandler(db, files.NewServiceWithIngest(db, files.IngestDeps{Storage: svc}), svc, nil, nil, nil, nil, nil, nil)
 
 	fix := seedLibrary(t, db)
 	fid := createFile(t, db, fix.LibraryID, fix.UserID, "format.jpg", true, nil)
